requirements.txt
Pillow>=9.0.0
argparse>=1.4.0
